10/07/2024

+++ src/main/java/com/example/HRApplication/Repositories/AttendanceRepository.java
@@ -11,8 +11,8 @@ import java.util.List;
 @Repository
 public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
     List<Attendance> findByUser(User user);
-    List<Attendance> findByUserId(Long userId);
+    List<Attendance> findByUserId(Integer userId);
     List<Attendance> findByDate(LocalDate date);
-    List<Attendance> findByUserIdAndDateBetween(Long userId, LocalDate startDate, LocalDate endDate);
+    List<Attendance> findByUserIdAndDateBetween(Integer userId, LocalDate startDate, LocalDate endDate);
     List<Attendance> findByDateBetween(LocalDate startDate, LocalDate endDate);
 }
